move new interface frmo CVXcanon to Solver
